+++ test/spectrumAlert/create.js
@@ -189,5 +189,21 @@ describe("spectrumAlert", function() {
                     );
     });
 
+    it("return error from spotUtil.getTokenAndConfigs", function(done){
+      var context = {
+        done: ()=>{
+          done()
+      }}
+
+      create.handler(
+        _.merge({
+          id:           'sig-11111111',
+        }, spectrumAlertConfig),
+        context
+      );
+    })
+
+
+
   });
 });