+++ test/subscription/create.js
@@ -93,5 +93,20 @@ describe("subscription", function() {
                     );
     });
 
+    it("return error from spotUtil.getTokenAndConfigs", function(done){
+      var context = {
+        done: ()=>{
+          done()
+      }}
+
+      create.handler(
+        _.merge({
+          id:           'sig-11111111',
+        }, subscriptionConfig),
+        context
+      );
+    })
+
+
   });
 });