+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.4.24] - 2018-09-21
+### Changed
+ - moved the `createGroup` function to `spotUtil`
+
+### Added
+ - `importAsg` accepts new field `groupConfig` for create and update
+ - `importBeanstalk` accepts new field `groupConfig` for create and update
+
 ## [0.4.23] - 2018-08-30
 ### Added
  - `importAsg` resource to create, delete and update